backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sse-starlette==2.0.0

# LangChain - Compatible versions (all use langchain-core<0.2.0)
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.5

# Web Search & Scraping

httpx==0.26.0  # Used for Google Custom Search API calls
beautifulsoup4==4.12.3
html2text==2024.2.26

# Database
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25

# PDF Export
markdown==3.5.2
weasyprint==60.2

# Academic Search
arxiv==2.1.0

# Rate Limiting & Caching
slowapi==0.1.9
cachetools==5.3.2

# Authentication (Clerk JWT verification)
PyJWT==2.8.0
cryptography==42.0.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async support
aiohttp==3.9.1
asyncio-throttle==1.0.2
